Pull in masonry to make it look good
This pulls in masonry to layout the images in a not-so-boring way. Also
update about page and make a first pass at making the site responsive.

Still a lot of cleanup that should be done. Currently the thumbnail
sizes are hardcoded which should be fixed but this works for now.